requirement.txt
asgiref==3.6.0
Django==4.1.5
django-cleanup==6.0.0
django-crispy-forms==1.14.0
fpdf==1.7.2
Pillow==9.4.0
sqlparse==0.4.3
